# docs/openapi/v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/residents": {
      "get": {
        "operationId": "ResidentsController_list",
        "parameters": [
          {
            "name": "unitId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 160,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResidentsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Residents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ResidentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResidentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Residents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/residents/{id}": {
      "patch": {
        "operationId": "ResidentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResidentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resident"
                }
              }
            }
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Residents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ResidentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Residents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "SIGRA API",
    "description": "Official SIGRA Phase 0 HTTP contract",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "ResidentUnitResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "parkingSpaces": {
            "type": "number"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "code",
          "address",
          "parkingSpaces",
          "active",
          "createdAt",
          "updatedAt"
        ]
      },
      "ResidentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          },
          "unitId": {
            "type": "string",
            "format": "uuid"
          },
          "unit": {
            "$ref": "#/components/schemas/ResidentUnitResponseDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "active",
          "unitId",
          "unit",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedResidentsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResidentResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "CreateResidentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "maxLength": 40
          },
          "unitId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72
          }
        },
        "required": [
          "name",
          "unitId",
          "email",
          "password"
        ]
      },
      "UpdateResidentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "maxLength": 40
          },
          "unitId": {
            "type": "string",
            "format": "uuid"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "ResidentialUnit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "parkingSpaces": {
            "type": "number"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "address",
          "parkingSpaces",
          "active",
          "createdAt",
          "updatedAt"
        ]
      },
      "Resident": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          },
          "unitId": {
            "type": "string"
          },
          "unit": {
            "$ref": "#/components/schemas/ResidentialUnit"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "active",
          "unitId",
          "unit",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Phase0Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "BAD_REQUEST",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "RATE_LIMITED",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "requestId": {
            "type": "string"
          }
        }
      }
    }
  }
}